refactor: pc/spの判別用pluginを追加、URLを集約
